requirements.txt
requests==2.32.4
python-dotenv==1.1.0
langchain==0.3.25
langchain-anthropic==0.3.15
langchain-core==0.3.65
langgraph==0.4.8
langchain-mcp-adapters>=0.1.0

# MCP dependencies
mcp>=1.9.0
mcp-server-fetch>=2025.4.0
httpx>=0.27.2
httpx-sse>=0.4.0
pydantic-settings>=2.5.2
python-multipart>=0.0.9
sse-starlette>=1.6.1
starlette>=0.27.0
uvicorn>=0.23.1

# Additional dependencies for MCP examples
markdownify>=0.13.1
protego>=0.3.1
readabilipy>=0.2.0
colorama>=0.4.6